// src/app/api/models/jogo.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Jogo {
  classificacaoEtaria?: string;
  dataLancamento?: string;
  desenvolvedora?: string;
  genero?: string;
  id?: number;
  multiplayer?: boolean;
  plataforma?: string;
  titulo?: string;
}